requirements.txt
streamlit
langchain
langchain_community
langchain_google_genai
google-generativeai
faiss-cpu
python-dotenv
PyPDF2
typing